requirements.txt
fastapi
gradio
pandas
scikit-learn
matplotlib
python-dotenv
google-generativeai
uvicorn